Ajustes no projeto

+++ src/main/java/br/com/idosos/servico/MedicamentoServico.java
@@ -3,6 +3,10 @@ package br.com.idosos.servico;
 import java.util.List;
 import java.util.Optional;
 
+import br.com.idosos.excessoes.MedicamentoNaoEncontrado;
+import br.com.idosos.excessoes.PacienteNaoEncontrado;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import br.com.idosos.modelo.Medicamento;
@@ -17,27 +21,47 @@ public class MedicamentoServico {
 	private final PacienteRepositorio pacienteRepositorio;
 	private final MedicamentoRepositorio medicamentoRepositorio;
 	
-	public Medicamento cadastrarMedicamento(Medicamento medicamento,Long pacienteId) {
-		var paciente = pacienteRepositorio.findById(pacienteId).get();
+	public Medicamento cadastrarMedicamento(Long pacienteId, Medicamento medicamento) {
+		var paciente = pacienteRepositorio.findById(pacienteId).orElseThrow(PacienteNaoEncontrado::new);
 		medicamento.setPaciente(paciente);
 		return medicamentoRepositorio.save(medicamento);
 	}
 	
-	public List<Medicamento>listarMedicamentos(){
-		return medicamentoRepositorio.findAll();
+	public Page<Medicamento> listarMedicamentos(Pageable pageable){
+		return medicamentoRepositorio.findAll(pageable);
 	}
 	
 	public Medicamento buscarPorId(Long id) {
-		Optional<Medicamento>buscar = medicamentoRepositorio.findById(id);
-		return buscar.get();
+	return  medicamentoRepositorio.findById(id).orElseThrow(()-> new MedicamentoNaoEncontrado("Medicamento não encontrado."));
+
 	}
 	
-	public Medicamento atualizarMedicamento(Medicamento medicamento,Long id) {
-		medicamento.setId(id);
-		return medicamentoRepositorio.save(medicamento);
+	public Medicamento atualizarMedicamento(Long id, Medicamento medicamento) {
+		var buscarMed =  medicamentoRepositorio.findById(id).orElseThrow(PacienteNaoEncontrado::new);
+		if(medicamento.getNomeMedicamento() != null){
+			buscarMed.setNomeMedicamento(medicamento.getNomeMedicamento());
+		}
+        if(medicamento.getVia() != null){
+			buscarMed.setVia(medicamento.getVia());
+		}
+		if (medicamento.getEnfermidade() != null){
+			buscarMed.setEnfermidade(medicamento.getEnfermidade());
+		}
+		if(medicamento.getDescricao() != null){
+			buscarMed.setDescricao(medicamento.getDescricao());
+		}
+
+		return medicamentoRepositorio.save(buscarMed);
+
+
 	}
     
 	public void excluirMedicamento(Long id) {
-		medicamentoRepositorio.deleteById(id);
+
+		if(!medicamentoRepositorio.existsById(id)){
+			throw new MedicamentoNaoEncontrado();
+		}else{
+			medicamentoRepositorio.deleteById(id);
+		}
 	}
 }
